blip-server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
transformers
torch
Pillow
opencv-python
aiortc
av
numpy
python-socketio